+++ src/Ergo.Compiler.Emission/Emitter/Emitter.cs
@@ -123,7 +123,7 @@ public class Emitter
             if (!kb.TryResolve(sign, out var label)) {
                 // Check if it's declared dynamic — emit call for runtime resolution
                 var c = ctx.Constant(sign.Functor.Value);
-                var dynSig = (Signature)(c, sign.Arity.TryGetValue(out var dynA) ? dynA : (int)Signature.VARIADIC);
+                var dynSig = (Signature)(c, sign.Arity.TryGetValue(out var dynA) ? dynA : Signature.VARIADIC);
                 if (kb.Labels.ContainsKey(dynSig))
                     label = dynSig;
                 else
@@ -137,7 +137,7 @@ public class Emitter
     protected virtual void Predicate(EmitterContext ctx, Predicate predicate)
     {
         var p = ctx.Constant(predicate.Signature.Functor.Value);
-        var n = predicate.Signature.Arity.TryGetValue(out var predArity) ? predArity : (int)Signature.VARIADIC;
+        var n = predicate.Signature.Arity.TryGetValue(out var predArity) ? predArity : Signature.VARIADIC;
         ctx.Label((p, n), ctx.PC);
         var clauseCtxs = new EmitterContext[predicate.Clauses.Count];
 
@@ -197,7 +197,7 @@ public class Emitter
                 break;
             case StaticGoal @static:
                 var p1 = ctx.Constant(@static.Callee.Signature.Functor.Value);
-                var n1 = @static.Callee.Signature.Arity.TryGetValue(out var sA) ? sA : (int)Signature.VARIADIC;
+                var n1 = @static.Callee.Signature.Arity.TryGetValue(out var sA) ? sA : Signature.VARIADIC;
                 ctx.Emit(call((Signature)(p1, n1)));
                 break;
             case DynamicGoal @dynamic:
@@ -461,7 +461,7 @@ public class Emitter
                     Write(sc, goalArgs, k, vars, deep: true);
                 var sig = goal.GetSignature().GetOrThrow();
                 var p = sc.Constant(sig.Functor.Value);
-                sc.Emit(Ops.call((Signature)(p, sig.Arity.TryGetValue(out var gA) ? gA : (int)Signature.VARIADIC)));
+                sc.Emit(Ops.call((Signature)(p, sig.Arity.TryGetValue(out var gA) ? gA : Signature.VARIADIC)));
             }
         }
